requirements.txt
Pillow>=10.0.0
exifread>=3.0.0
piexif>=1.1.3